moved hypergeom calculation logic to util (also fixes lint error)
